Move ui files into a ui folder
