+++ src/ui/components/SplashScreen.vue
