+++ src/ui/components/shared/LogLines.vue
