+++ src/ui/components/shared/log/AccountCardStatusIndicator.vue
