+++ src/ui/components/shared/log/LogViewer.vue
@@ -25,8 +25,8 @@
 
 <script lang="ts">
 import { defineComponent, PropType } from '@vue/composition-api';
-import { AccountsState } from '@/components/app/accountsState';
-import AccountCard from '@/components/shared/log/AccountCard.vue';
+import { AccountsState } from '@/ui/components/app/accountsState';
+import AccountCard from '@/ui/components/shared/log/AccountCard.vue';
 import { Levels } from './types';
 
 const levelToClass = {
