+++ src/ui/plugins/vuetify.js
