+++ src/ui/store/plugins/persisted-config.ts
