+++ test/unit/specs/importers.vue.spec.ts
@@ -1,7 +1,7 @@
 import { createLocalVue, shallowMount, Wrapper } from '@vue/test-utils';
 import Vuex, { Store } from 'vuex';
-import Importers from '../../../src/components/app/Importers.vue';
-import AddImporter from '../../../src/components/app/importers/AddImporter.vue';
+import Importers from '../../../src/ui/components/app/Importers.vue';
+import AddImporter from '../../../src/ui/components/app/importers/AddImporter.vue';
 import { inputVendors } from '../../../src/originalBudgetTrackingApp';
 
 const localVue = createLocalVue();
